+++ pypump/models/activity.py
@@ -66,7 +66,7 @@ class Activity(PumpObject):
                 data["object"][key] = data[key]
 
         Mapper(pypump=self._pump).parse_map(self, data=data)
-        self.add_links(data)
+        self._add_links(data)
 
         return self