+++ pypump/models/feed.py
@@ -267,7 +267,7 @@ class Feed(PumpObject):
 
     def unserialize(self, data={}):
         Mapper(pypump=self._pump).parse_map(self, data=data)
-        self.add_links(data)
+        self._add_links(data)
         self.url = data.get('pump_io', {}).get('proxyURL') or self.url
         return self